Change implementation of OPEN list
In the previous implementation when a node was expanded its neighbors
would always be added to the OPEN list because the PriorityQueue
did not check if the neighbor's location was already in the list.
The PrioritySet implementation checks if a location is already in the
set and only adds the item if it has a lower rating than the existing
item.